server/requirements.txt
Flask
Flask-CORS
APScheduler
python-dotenv
SQLAlchemy
Flask-SQLAlchemy
requests


